feat(formatting): Implement auto-reformatting of input script
This adds an editor action to trigger automatic reformatting of the
input according to my personal stylistic preferences for this language.
Whitespace between tokens is replaced with the respectively correct
amount and sort (i.e., a single space or a single newline).
In case where the separator is a newline and the input already contained
multiple newlines (i.e., one or more empty lines between statements),
one empty line is kept.
Comments are also handled correctly.
The formatter is very forgiving and even handles abruptly ending input
gracefully, formatting everything up until the end of the script.