feat(export): Implement PDF export with PDFKit
This implements a custom renderer that outputs to a PDFKit PDFDocument.
Based on this, a PDF data stream can be generated to a Blob and
downloaded.
Webpack fallbacks needed to be added to support the browser, and since
I could not get PDFKit compilation to work, I unfortunately have to
import the pre-built "standalone" file.