feat: Use Vite as build tool and implement own BlobStream (#38)
Vite can build this project directly without any configuration, which is
pretty awesome. It allows us to save tons of dependencies compared to
Webpack (and it's much faster, too).

Only the 'blob-stream' dependency was incompatible with Vite, so this
patch includes a custom implementation that doesn't depend on any Node
internals and hence works nicely in the browser.

Due to this, a few more shims (buffer, process, stream-browserify, util)
could be removed as well. 'eventemitter3' was replaced with 'events' to
avoid incompatibilities with Node's own EventEmitter interface.